% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_module.R
\name{analyze_ehr_data}
\alias{analyze_ehr_data}
\title{Analyze ehr data}
\usage{
analyze_ehr_data(path)
}
\arguments{
\item{path}{path to dataset}
}
\value{
a summary staatistics and plots
}
\description{
Analyze ehr data
}
\examples{
analyze_ehr_data("~/data/ehr.csv")
}
