% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_module.R
\name{visualize_ehr_data}
\alias{visualize_ehr_data}
\title{Visualize ehr data}
\usage{
visualize_ehr_data(ehr_data)
}
\arguments{
\item{ehr_data}{}
}
\value{
plots indicating the relationship between age and diseases
}
\description{
Visualize ehr data
}
\examples{
visualize_ehr_data(ehr_data)
}
